requirements.txt
telethon>=1.28.5
python-dotenv>=1.0.0
tqdm>=4.66.1
pyyaml>=6.0
torch>=2.0.0
transformers>=4.30.0
bitsandbytes>=0.39.0
huggingface-hub>=0.16.4
nltk>=3.8.1
pillow>=10.0.0
numpy>=1.24.0

